few changes

+++ test.py
@@ -5,16 +5,19 @@ import sys
 size = int(sys.argv[1])
 dir = '/mnt/disk1/tan_hm/corpus/train'
 files = [os.path.join(dir, i) for i in os.listdir(dir)]
+if os.path.isfile('/mnt/disk1/tan_hm/test_pkl.pkl'):
+    with open('/mnt/disk1/tan_hm/test_pkl.pkl', 'rb') as f:
+        res = pickle.load(f)
+else:
+    counter = 0
+    seg = 2**30
+    res = []
 
-counter = 0
-seg = 2**30
-res = []
+    while size > 0:
+        with open(files[counter]) as f:
+            res.append(f.read(seg))
+        counter += 1
+        size -= 1
 
-while size > 0:
-    with open(files[counter]) as f:
-        res.append(f.read(seg))
-    counter += 1
-    size -= 1
-
-with open('/mnt/disk1/tan_hm/test_pkl.pkl', 'wb') as f:
-    pickle.dump(res, f, protocol=pickle.HIGHEST_PROTOCOL)
+    with open('/mnt/disk1/tan_hm/test_pkl.pkl', 'wb') as f:
+        pickle.dump(res, f, protocol=pickle.HIGHEST_PROTOCOL)
